Feat(CH-47,73,49): Estrutura básica da tela de seleção, estilização com base no design validado e implementação do layout responsivo
