+++ package.json
@@ -15,6 +15,7 @@
     "postcss": "^8.5.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.5.1",
     "tailwindcss": "^3.4.17"
   },
   "devDependencies": {
